fix: func check demmaped bits

+++ pluto/dev/include/sharedData.hpp
@@ -17,7 +17,7 @@ struct sharedData
     std::vector<std::complex<float>> rx_complex_fft_gui;
     std::vector<int16_t> tx_buffer;
     std::vector<float> demaped_bits;
-    std::vector<int> bits;
+    std::vector<float> bits;
     std::vector<float> shifted_magnitude;
     std::vector<float> argument;
     std::vector<float> frequency_axis;
